icm42688: RTIO device for streaming
An experimental idea to enable streaming fifo data from the sensor using
RTIO with the existing sensor API. A sensor reader could easily be built on
top of the buffer to produce meaningful values.

There are many ways this can be built currently. Using either the existing
trigger work context (system workqueue or own thread) when SPI_RTIO=n or
using a sequence. With or without SPI_SAM_DMA (where only polling SPI
transfers are used).

This is a really nice setup to explore and experiment with RTIO, SPI, and
a really fast sensor.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>