samples: tdk_robokit1: Poll the magnetometer
While an asynchronous transfer is on going for the 6 axis imu, concurrently
read the magnetometer with its blocking fetch call.

In the future it'd be nice to have the magnetometer also work with RTIO
driven by a timer. In fact having all the sensors read on a timer would be
incredibly nice to showcase.

Signed-off-by: Tom Burdick <thomas.burdick@intel.com>